--- commonlib/commonlib.aar
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:b42781c1c6690bfde21f3857fc71ef20de057723006d44cf86753080e9cee45a
-size 3316604
