+++ course/src/main/java/in/testpress/course/ui/MyCoursesFragment.java
@@ -1,13 +1,6 @@
 package in.testpress.course.ui;
 
 import android.os.Bundle;
-import android.util.Log;
-import android.view.Menu;
-import android.view.MenuInflater;
-import android.view.MenuItem;
-import android.view.View;
-
-import androidx.annotation.NonNull;
 import androidx.fragment.app.FragmentActivity;
 import androidx.loader.content.Loader;
 
@@ -18,11 +11,8 @@ import java.util.List;
 
 import in.testpress.core.TestpressException;
 import in.testpress.core.TestpressSDKDatabase;
-import in.testpress.core.TestpressSdk;
-import in.testpress.core.TestpressSession;
 import in.testpress.course.R;
 import in.testpress.course.helpers.CourseLastSyncedDate;
-import in.testpress.models.InstituteSettings;
 import in.testpress.models.greendao.Course;
 import in.testpress.models.greendao.CourseDao;
 import in.testpress.course.pagers.CoursePager;
@@ -36,7 +26,6 @@ public class MyCoursesFragment extends BaseDataBaseFragment<Course, Long> {
     private TestpressCourseApiClient mApiClient;
     private CourseDao courseDao;
     private ArrayList<String> tags = new ArrayList<String>();
-    private InstituteSettings instituteSettings;
 
     public static void show(FragmentActivity activity, int containerViewId) {
         activity.getSupportFragmentManager().beginTransaction()
@@ -49,14 +38,7 @@ public class MyCoursesFragment extends BaseDataBaseFragment<Course, Long> {
         super.onCreate(savedInstanceState);
         mApiClient = new TestpressCourseApiClient(getActivity());
         courseDao = TestpressSDKDatabase.getCourseDao(getActivity());
-        instituteSettings = TestpressSdk.getTestpressSession(requireContext()).getInstituteSettings();
         initializeTags();
-    }
-
-    @Override
-    public void onViewCreated(View view, Bundle savedInstanceState) {
-        super.onViewCreated(view, savedInstanceState);
-        setHasOptionsMenu(isCustomTestGenerationEnabled());
     }
 
     private void initializeTags() {
@@ -162,47 +144,4 @@ public class MyCoursesFragment extends BaseDataBaseFragment<Course, Long> {
         setEmptyText(R.string.testpress_no_courses, R.string.testpress_no_courses_description,
                     R.drawable.ic_error_outline_black_18dp);
     }
-
-    private boolean isCustomTestGenerationEnabled() {
-        //return instituteSettings.getAllowCustomTestGeneration() != null && course.getAllowCustomTestGeneration();
-        return true;
-    }
-
-    @Override
-    public void onCreateOptionsMenu(@NonNull Menu menu, @NonNull MenuInflater inflater) {
-        inflater.inflate(R.menu.custom_test_generation, menu);
-        super.onCreateOptionsMenu(menu, inflater);
-    }
-
-    @Override
-    public boolean onOptionsItemSelected(@NonNull MenuItem item) {
-        if (item.getItemId() == R.id.custom_test_icon) {
-            openCustomTestGenerationActivity();
-            return true;
-        }
-        return super.onOptionsItemSelected(item);
-    }
-
-    private void openCustomTestGenerationActivity() {
-        startActivity(
-                CustomTestGenerationActivity.Companion.createIntent(
-                        requireContext(),
-                        "Custom Module",
-                        "/courses/custom_test_generation/?"+constrictQueryParamForAvailableCourses(),
-                        true,
-                        CustomTestGenerationActivity.class
-                )
-        );
-    }
-
-    private String constrictQueryParamForAvailableCourses(){
-        Log.d("TAG", getCourses().size()+"");
-        StringBuilder queryParam = new StringBuilder();
-        for (Course course : getCourses()) {
-            queryParam.append("course_id=").append(course.getId()).append("%26");
-        }
-        Log.d("TAG", queryParam.toString());
-        return queryParam.toString();
-    }
-
 }
