+++ exam/src/main/java/in/testpress/exam/ui/TestActivity.java
@@ -9,7 +9,6 @@ import androidx.annotation.StringRes;
 import androidx.loader.app.LoaderManager;
 import androidx.loader.content.Loader;
 
-import android.util.Log;
 import android.view.View;
 import android.widget.Button;
 import android.widget.LinearLayout;
