--- mobilertc/mobilertc.aar
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:59b44e1d897db527c98fdad97de29644415422be0e232c60174124f050c03338
-size 212668096
